pyproject.toml
[tool.poetry]
name = "transactionapp"
version = "0.1.0"
description = ""
authors = ["Timur Gorashchenko <timyr98239718@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flake8 = "^7.3.0"
black = "^25.9.0"
mypy = "^1.18.2"
isort = "^6.0.1"
pandas = "^2.3.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [
    'venv',
    'tests'
]

[tool.isort]
line_length = 119
